style: Revamp the whole site for mobile devices.
Followed design technique of starting small. The whole website was revamped to fit on a small viewport such as a mobile phone.

js\restaurant_info.js was updated to provide classes for the different parts of the reviews in order to style them effectively.

Both HTML files were updated with meta tags, and a font link.

CSS file updated to work with a viewport set at 360px.